+++ CHANGELOG.md
@@ -3,9 +3,12 @@ Dependency-free, type-safe Dart unit conversion library. Decouple specific unit 
 
 ## [Unreleased]
 ### Added
-- Update toString methods for a few measurement types
-- Support for Quantity measurements (e.g. moles)
+- Support for quantity measurements (e.g. moles)
+- Support for electric charge measurements (e.g. amperes)
 - Support for daltons and electron rest mass
+
+### Changed
+- Update toString methods for a few measurement types
 
 ## [1.2.0] - 2021-03-25
 Note: a lot of the old methods are marked "deprecated" as of 1.2.0 - these methods will remain in the library